Moving functions and tests into more sensible file locations
