+++ requests_app/management/commands/_insert_ci.py
@@ -8,8 +8,7 @@ from django.db import transaction
 from django.db.models import QuerySet
 
 from ._utils import sortable_version, normalize_version
-from ._insert_panel import flag_active_links_for_ci, \
-    make_provisional_ci_panel_link
+from ._insert_panel import make_provisional_ci_panel_link
 
 from requests_app.models import (
     Panel,
@@ -25,6 +24,44 @@ from requests_app.models import (
     ClinicalIndicationTestMethodHistory,
     PanelGeneHistory,
 )
+
+
+def _flag_active_links_for_ci(r_code: str, user: str) \
+    -> QuerySet[ClinicalIndicationPanel] | None:
+    """
+    Controller function which takes a clinical indication r code, and flags ACTIVE links between the CI 
+    and its panels for manual review.
+    This is useful when a new CI is added, e.g. from test directory, and the user might want to switch to 
+    using that for a panel instead.
+    Note that a ClinicalIndication might have multiple CI-Panel links!
+    """
+    ci_panel_instances: QuerySet[
+        ClinicalIndicationPanel
+    ] = ClinicalIndicationPanel.objects.filter(
+        clinical_indication__r_code=r_code,
+        current=True,  # get those that are active
+    )
+
+    # for each previous CI-Panel instance, flag for manual review, and add to history
+    if ci_panel_instances:
+        for ci_panel_instance in ci_panel_instances:
+            ci_panel_instance.needs_review = True
+            ci_panel_instance.save()
+
+            ClinicalIndicationPanelHistory.objects.create(
+                    clinical_indication_panel=ci_panel_instance,
+                    note="Flagged for manual review - new clinical indication provided",
+                    user=user
+                )
+
+            print(
+                'Flagged for manual review - a new clinical indication is available'
+            )
+
+        return ci_panel_instances
+
+    else:
+        return None
 
 
 def _get_td_version(filename: str) -> str | None:
@@ -334,9 +371,7 @@ def insert_test_directory_data(json_data: dict, user:str, force: bool = False) -
 
             for previous_ci in previous_cis:
                 previous_panel_ci_links = \
-                    flag_active_links_for_ci(previous_ci.pk, 
-                                            previous_ci.name,
-                                            user)
+                    _flag_active_links_for_ci(previous_ci.r_code, user)
                 if previous_panel_ci_links:
                     make_provisional_ci_panel_link(previous_panel_ci_links, ci_instance, user, "ci")
